Add export/import commands for pass
This should allow for exporting a password store using a different age
key. Useful for if you want to export a password store to a separate
device, such as a laptop. You would ideally have a separate private key
for that laptop, in which case that would be used for handling the
storing of passwords.